+++ .sqlx/query-912feea95af77d642446aa5b44f8ba93bc652c1af359e8c58848575f082fcc4b.json
@@ -1,0 +1,31 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "INSERT INTO emote_usage (message_id, emote_id, user_id, channel_id, guild_id,\n             used_at, usage_type) VALUES ($1, $2, $3, $4, $5, $6, $7)",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Int8",
+        "Int8",
+        "Int8",
+        "Int8",
+        "Int8",
+        "Int8",
+        {
+          "Custom": {
+            "name": "emoteusagetype",
+            "kind": {
+              "Enum": [
+                "Message",
+                "ReactionAdd",
+                "ReactionRemove"
+              ]
+            }
+          }
+        }
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "912feea95af77d642446aa5b44f8ba93bc652c1af359e8c58848575f082fcc4b"
+}
