+++ .sqlx/query-9c4bd1d57888bd6a881182c3e41aa5c29e75ad866c3370a55effaf1ed03c99c0.json
@@ -1,0 +1,15 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "INSERT INTO emotes (emote_name, discord_id) VALUES ($1, $2) ON CONFLICT (discord_id) DO NOTHING",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Text",
+        "Int8"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "9c4bd1d57888bd6a881182c3e41aa5c29e75ad866c3370a55effaf1ed03c99c0"
+}
