+++ .sqlx/query-b14788c976db92f360608f42ebdb4cd66e3776d50ad8f96a4fe94359ae824982.json
@@ -1,0 +1,18 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "INSERT INTO sticker_usage (message_id, user_id, channel_id, guild_id, sticker_id) VALUES ($1, $2, $3, $4, $5)",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Int8",
+        "Int8",
+        "Int8",
+        "Int8",
+        "Int8"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "b14788c976db92f360608f42ebdb4cd66e3776d50ad8f96a4fe94359ae824982"
+}
