+++ .sqlx/query-c820d35c389dd7cfee63d030ef47915714a0dcaf741b5b9ecc5e8c55e858b58a.json
@@ -1,0 +1,15 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "INSERT INTO stickers (sticker_id, sticker_name) VALUES ($1, $2) ON CONFLICT (sticker_id) DO NOTHING",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Int8",
+        "Text"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "c820d35c389dd7cfee63d030ef47915714a0dcaf741b5b9ecc5e8c55e858b58a"
+}
