+++ jamespy_events/src/handlers/messages/mod.rs
@@ -1,51 +1,21 @@
 use std::collections::{HashMap, HashSet};
 use std::fmt::Write;
-use std::ops::Deref;
 use std::sync::Arc;
 use std::time::Instant;
+
+mod database;
 
 use crate::helper::{get_channel_name, get_guild_name, get_guild_name_override};
 use crate::{Data, Error};
 
 use ::serenity::all::{CreateEmbed, CreateMessage, GetMessages};
-use chrono::Utc;
+use database::{insert_deletion, insert_edit, insert_message};
 use jamespy_data::structs::{Decay, InnerCache};
 use poise::serenity_prelude::{
     self as serenity, ChannelId, Colour, CreateEmbedFooter, GuildId, Message, MessageId,
     MessageUpdateEvent, UserId,
 };
 use small_fixed_array::FixedString;
-use sqlx::postgres::{PgArgumentBuffer, PgTypeInfo};
-use sqlx::{query, Encode, Postgres, Transaction, Type};
-
-use small_fixed_array::ValidLength;
-
-// Foreign trait foreign type stuff.
-pub struct FuckRustRules<'a, LenT: ValidLength>(&'a FixedString<LenT>);
-
-impl<LenT: ValidLength> Deref for FuckRustRules<'_, LenT> {
-    type Target = FixedString<LenT>;
-
-    fn deref(&self) -> &Self::Target {
-        self.0
-    }
-}
-
-impl<LenT: ValidLength> Type<Postgres> for FuckRustRules<'_, LenT> {
-    fn type_info() -> PgTypeInfo {
-        <&str as Type<Postgres>>::type_info()
-    }
-
-    fn compatible(ty: &PgTypeInfo) -> bool {
-        <&str as Type<Postgres>>::compatible(ty)
-    }
-}
-
-impl<LenT: ValidLength> Encode<'_, Postgres> for FuckRustRules<'_, LenT> {
-    fn encode_by_ref(&self, buf: &mut PgArgumentBuffer) -> sqlx::encode::IsNull {
-        <&str as Encode<Postgres>>::encode(self.as_str(), buf)
-    }
-}
 
 pub async fn message(ctx: &serenity::Context, msg: &Message, data: Arc<Data>) -> Result<(), Error> {
     let (flagged_words, patterns) = {
@@ -117,89 +87,6 @@ pub async fn message(ctx: &serenity::Context, msg: &Message, data: Arc<Data>) ->
     Ok(())
 }
 
-async fn insert_message(
-    mut transaction: Transaction<'_, Postgres>,
-    message: &Message,
-) -> Result<(), Error> {
-    let guild_id = message.guild_id.map(|g| g.get() as i64);
-    let channel_id = message.channel_id.get() as i64;
-    let user_id = message.author.id.get() as i64;
-    let message_id = message.id.get() as i64;
-
-    query!(
-        "INSERT INTO channels (channel_id, guild_id)
-         VALUES ($1, $2)
-         ON CONFLICT (channel_id) DO NOTHING",
-        channel_id,
-        guild_id
-    )
-    .execute(&mut *transaction)
-    .await?;
-
-    if let Some(guild_id) = guild_id {
-        query!(
-            "INSERT INTO guilds (guild_id)
-             VALUES ($1)
-             ON CONFLICT (guild_id) DO NOTHING",
-            guild_id
-        )
-        .execute(&mut *transaction)
-        .await?;
-    }
-
-    query!(
-        "INSERT INTO users (user_id)
-         VALUES ($1)
-         ON CONFLICT (user_id) DO NOTHING",
-        user_id
-    )
-    .execute(&mut *transaction)
-    .await?;
-
-    query!(
-        "INSERT INTO messages (message_id, guild_id, channel_id, user_id, content, created_at)
-         VALUES ($1, $2, $3, $4, $5, $6)",
-        message_id,
-        guild_id,
-        channel_id,
-        user_id,
-        &FuckRustRules(&message.content),
-        message.id.created_at().unix_timestamp()
-    )
-    .execute(&mut *transaction)
-    .await?;
-
-    if !message.embeds.is_empty() {
-        query!(
-            "INSERT INTO embeds (message_id, embed_data)
-             VALUES ($1, $2)
-             ON CONFLICT (message_id) DO NOTHING",
-            message_id,
-            serde_json::to_value(message.embeds.clone())?
-        )
-        .execute(&mut *transaction)
-        .await?;
-    }
-
-    for attachment in &message.attachments {
-        query!(
-            "INSERT INTO attachments (attachment_id, message_id, file_name, file_size, file_url)
-             VALUES ($1, $2, $3, $4, $5)",
-            attachment.id.get() as i64,
-            message_id,
-            &FuckRustRules(&attachment.filename),
-            attachment.size as i32,
-            &FuckRustRules(&attachment.url)
-        )
-        .execute(&mut *transaction)
-        .await?;
-    }
-
-    transaction.commit().await?;
-
-    Ok(())
-}
-
 pub async fn message_edit(
     ctx: &serenity::Context,
     old_if_available: &Option<Message>,
@@ -253,67 +140,6 @@ pub async fn message_edit(
         }
         _ => {}
     }
-    Ok(())
-}
-
-async fn insert_edit(
-    mut transaction: Transaction<'_, Postgres>,
-    message: &Message,
-) -> Result<(), Error> {
-    let guild_id = message.guild_id.map(|g| g.get() as i64);
-    let channel_id = message.channel_id.get() as i64;
-    let user_id = message.author.id.get() as i64;
-    let message_id = message.id.get() as i64;
-
-    query!(
-        "INSERT INTO channels (channel_id, guild_id)
-         VALUES ($1, $2)
-         ON CONFLICT (channel_id) DO NOTHING",
-        channel_id,
-        guild_id
-    )
-    .execute(&mut *transaction)
-    .await?;
-
-    if let Some(guild_id) = guild_id {
-        query!(
-            "INSERT INTO guilds (guild_id)
-             VALUES ($1)
-             ON CONFLICT (guild_id) DO NOTHING",
-            guild_id
-        )
-        .execute(&mut *transaction)
-        .await?;
-    }
-
-    query!(
-        "INSERT INTO users (user_id)
-         VALUES ($1)
-         ON CONFLICT (user_id) DO NOTHING",
-        user_id
-    )
-    .execute(&mut *transaction)
-    .await?;
-
-    let timestamp = message
-        .edited_timestamp
-        .map_or_else(|| Utc::now().timestamp(), |t| t.unix_timestamp());
-
-    query!(
-        "INSERT INTO message_edits (message_id, channel_id, guild_id, user_id, content, \
-         edited_at) VALUES ($1, $2, $3, $4, $5, $6)",
-        message_id,
-        channel_id,
-        guild_id,
-        user_id,
-        &FuckRustRules(&message.content),
-        timestamp
-    )
-    .execute(&mut *transaction)
-    .await?;
-
-    transaction.commit().await?;
-
     Ok(())
 }
 
@@ -478,65 +304,6 @@ pub async fn message_delete(
             }
         }
     }
-    Ok(())
-}
-
-async fn insert_deletion(
-    mut transaction: Transaction<'_, Postgres>,
-    message: &Message,
-) -> Result<(), Error> {
-    let guild_id = message.guild_id.map(|g| g.get() as i64);
-    let channel_id = message.channel_id.get() as i64;
-    let user_id = message.author.id.get() as i64;
-    let message_id = message.id.get() as i64;
-
-    query!(
-        "INSERT INTO channels (channel_id, guild_id)
-         VALUES ($1, $2)
-         ON CONFLICT (channel_id) DO NOTHING",
-        channel_id,
-        guild_id
-    )
-    .execute(&mut *transaction)
-    .await?;
-
-    if let Some(guild_id) = guild_id {
-        query!(
-            "INSERT INTO guilds (guild_id)
-             VALUES ($1)
-             ON CONFLICT (guild_id) DO NOTHING",
-            guild_id
-        )
-        .execute(&mut *transaction)
-        .await?;
-    }
-
-    query!(
-        "INSERT INTO users (user_id)
-         VALUES ($1)
-         ON CONFLICT (user_id) DO NOTHING",
-        user_id
-    )
-    .execute(&mut *transaction)
-    .await?;
-
-    let timestamp = Utc::now().timestamp();
-
-    query!(
-        "INSERT INTO message_deletion (message_id, channel_id, guild_id, user_id, content, \
-         deleted_at) VALUES ($1, $2, $3, $4, $5, $6)",
-        message_id,
-        channel_id,
-        guild_id,
-        user_id,
-        &FuckRustRules(&message.content),
-        timestamp
-    )
-    .execute(&mut *transaction)
-    .await?;
-
-    transaction.commit().await?;
-
     Ok(())
 }
